pyproject.toml
[project]
name = "mcpcrunch"
version = "0.2.0"
description = "Comprehensive Security and Structural Validation Framework for the OpenMCP Specification"
authors = [
    { name = "Anish Chelliah CR", email = "anishchelliah.cr@google.com" }
]
license = { text = "MIT" }
dependencies = [
    "jsonschema>=4.20.0",
    "pydantic>=2.5.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
]
readme = "README.md"
requires-python = ">=3.10"

[project.scripts]
mcpcrunch = "mcpcrunch.cli:main"

[build-system]
requires = ["setuptools==69.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mcpcrunch = ["*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
collect_ignore_glob = ["src/*"]
